# Optimized example fences. Waited immediately. Now waits before reuse (next frame)