--- a/examples/src/bin/triangle.rs
+++ b/examples/src/bin/triangle.rs
@@ -394,6 +394,7 @@ fn main() {
             record_submit_commandbuffer(
                 &base.device,
                 base.draw_command_buffer,
+                base.draw_commands_reuse_fence,
                 base.present_queue,
                 &[vk::PipelineStageFlags::COLOR_ATTACHMENT_OUTPUT],
                 &[base.present_complete_semaphore],