The examples (triangle and texture) wait for a fence immediately after submitting a command buffer. This is not a good practice for performance. I moved the wait until the command buffer is reused (one frame later).

Measured average frame times on Intel i7-1065g7 (4K display, 200% scaling).

**Triangle sample:**
Old = 3.5ms
New = 2.4ms

**Texture sample:**
Old = 4.2ms
New = 3.0ms