@MaikKlein 2020-10-02
Thanks for the PR :), I'll have a look at it later today.